logger restructured
